[swagger/api]: main swagger yaml added and business logic for api moved from routes/index.js to api/controllers/api.js -refs #156
