+++ routes/index.js
@@ -273,131 +273,6 @@ exports.asn = function(req, res) {
   });
 };
 
-// api
-exports.apiRisk = function(req, res) {
-	if (Object.keys(req.params).length) {
-    logic.getRiskAPI(req.params).then(function(results){
-      res.json(results[0]);
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  } else {
-    logic.getRiskAPI(req.query).then(function(results){
-      res.json(results[0]);
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  }
-};
-
-exports.apiCountry = function(req, res) {
-  // handles with query and /{id} for REST
-  if (Object.keys(req.params).length) {
-    logic.getCountryAPI(req.params).then(function(results){
-      res.json(results[0]);
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  } else {
-    logic.getCountryAPI(req.query).then(function(results){
-      res.json(results[0]);
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  }
-};
-
-exports.apiAsn = function(req, res) {
-  // handles with query and /{asn} for REST
-  if (Object.keys(req.params).length) {
-    logic.getAsnAPI(req.params).then(function(results){
-      res.json(results[0]);
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  } else {
-    logic.getAsnAPI(req.query).then(function(results){
-      res.json(results[0]);
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  }
-};
-
-
-exports.apiCountByCountry = function(req, res) {
-	var queryOptions = {
-    country: req.query.country || "", risk: parseInt(req.query.risk) || -1,
-    start: req.query.start || "", end: req.query.end || "",
-    limit: checkLimit(req.query.limit) || "20",
-    page: req.query.page || 1,
-    granularity: req.query.granularity || 'week'
-  };
-  req.query.limit = checkLimit(queryOptions.limit);
-  queryOptions.country = queryOptions.country.toUpperCase();
-  queryOptions.offset = queryOptions.limit*(queryOptions.page-1);
-  try {
-    logic.getRowCountBYCountry(queryOptions).then(function(results) {
-      return results[0][0].count;
-    }).catch(function(err) {
-      res.json({error: err.message});
-    }).then(function (rows) {
-      logic.getCountByCountry(queryOptions).then(function(CountByCountry){
-        var normalized;
-        if (queryOptions.risk !== -1 && queryOptions.country && queryOptions.granularity === 'week') {
-          normalized = normalize(CountByCountry[0], '2017-01-02');
-        }
-        res.json({
-          total: rows,
-          results: normalized || CountByCountry[0]
-        });
-      }).catch(function(err) {
-        res.json({error: err.message});
-      });
-    });
-  } catch (err) {
-    res.json({error: err.message});
-  }
-};
-
-exports.apiCount = function(req, res) {
-  var queryOptions = {
-    country: req.query.country || "", risk: parseInt(req.query.risk) || -1, asn: parseInt(req.query.asn) || -1,
-    start: req.query.start || "", end: req.query.end || "",
-    limit:checkLimit(req.query.limit) || "20", page: req.query.page || 1
-  };
-  queryOptions.country = queryOptions.country.toLowerCase();
-  req.query.page = queryOptions.page;
-  req.query.limit = checkLimit(queryOptions.limit);
-
-  logic.getRowCount(queryOptions).then(function(results) {
-    return results[0][0].count;
-  }).catch(function(err) {
-    res.json({error: err.message});
-  }).then(function (rows){
-    queryOptions.offset = queryOptions.limit*(queryOptions.page-1);
-    var totalPages = Math.ceil(rows / queryOptions.limit);
-    var pages = checkCurrentPage(parseInt(queryOptions.page), totalPages, req.route.path, req.query);
-
-    logic.getTotalCount(queryOptions).then(function(results){
-      res.json({
-        status: "ok",
-        number_of_data_results: rows,
-        page: req.baseUrl+pages.curPage,
-        total_pages: totalPages,
-        next_page: req.baseUrl+pages.nextPage,
-        status_code: 200,
-        version: "1.2",
-        cached: false,
-        see_also: [],
-        time: new Date().toISOString(),
-        results: results[0]
-      });
-    }).catch(function(err) {
-      res.json({error: err.message});
-    });
-  });
-};
 
 exports.geo = function(req, res) {
   var geoJson = require('../data/geo.json');
